tests from single main
